[kamal] Commit

+++ C0777682_W2020_MID_TERM_MAD3004/C0777682_W2020_MID_TERM_MAD3004/calculateTotalBill.swift
@@ -10,5 +10,5 @@ import Foundation
 
 protocol CalculateTotalBill
 {
-   func  calculateTotalBill() -> Float
+   func  calculateTotalBill() -> Double
 }
